Revert "feat(env-check): pr comments"
This reverts commit 8586d40171302b3a10abdf78ef48ac7b952dd89a.
