+++ packages/environment-check/src/index.ts
@@ -1,3 +1,4 @@
 export * from './types';
 export * from './checks';
-export { convertResultsToMarkdown, storeResultsZip, archiveProject, convertResultsToString } from './output';
+export { convertResultsToMarkdown, storeResultsZip, archiveProject } from './output';
+export { formatResultsForClipboard } from './formatter';
