+++ packages/environment-check/src/output/index.ts
@@ -1,3 +1,2 @@
 export * from './markdown';
 export * from './zip';
-export * from './string';
